code refactor: move authentication logic from App Component to withAuthentication file; issue with login button on landingPage after logOut from App
