+++ src/js/components/Session/indexSession.js
@@ -1,3 +1,4 @@
 import AuthUserContext from "./context";
+import withAuthentication from './withAuthentication'
 
-export { AuthUserContext };
+export { AuthUserContext, withAuthentication };
